Use a base test class to avoid code duplication
