+++ sonar-ps-plugin/src/test/java/org/sonar/plugins/powershell/TokenizerSensorFileAnalysisTest.java
@@ -7,26 +7,17 @@ import java.util.Arrays;
 
 import org.apache.commons.io.FileUtils;
 import org.junit.Assert;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 import org.sonar.api.batch.fs.internal.DefaultInputFile;
 import org.sonar.api.batch.fs.internal.TestInputFileBuilder;
 import org.sonar.api.batch.sensor.internal.SensorContextTester;
-import org.sonar.api.utils.internal.JUnitTempFolder;
 import org.sonar.plugins.powershell.sensors.TokenizerSensor;
 
 @RunWith(Parameterized.class)
-public class TokenizerSensorFileAnalysisTest {
-
-	@Rule
-	public TemporaryFolder folder = new TemporaryFolder();
-
-	@org.junit.Rule
-	public JUnitTempFolder temp = new JUnitTempFolder();
+public class TokenizerSensorFileAnalysisTest extends BaseTest {
 
 	private final String key;
 
@@ -42,7 +33,7 @@ public class TokenizerSensorFileAnalysisTest {
 	@Test
 	public void testExecute() throws IOException {
 		SensorContextTester ctxTester = SensorContextTester.create(folder.getRoot().getAbsoluteFile().toPath());
-		ctxTester.settings().setProperty(Constants.PS_EXECUTABLE, "powershell.exe");
+		ctxTester.settings().setProperty(Constants.PS_EXECUTABLE, getPowerShellExecutable());
 
 		File baseFile = folder.newFile(key);
 		FileUtils.copyURLToFile(getClass().getResource("/testFiles/test.ps1"), baseFile);
