+++ sonar-ps-plugin/src/test/java/org/sonar/plugins/powershell/TokenizerSensorTest.java
@@ -6,26 +6,13 @@ import java.nio.file.Files;
 
 import org.apache.commons.io.FileUtils;
 import org.junit.Assert;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
 import org.sonar.api.batch.fs.internal.DefaultInputFile;
 import org.sonar.api.batch.fs.internal.TestInputFileBuilder;
 import org.sonar.api.batch.sensor.internal.SensorContextTester;
-import org.sonar.api.utils.internal.JUnitTempFolder;
 import org.sonar.plugins.powershell.sensors.TokenizerSensor;
 
-public class TokenizerSensorTest {
-
-	@Rule
-	public TemporaryFolder folder = new TemporaryFolder();
-
-	@org.junit.Rule
-	public JUnitTempFolder temp = new JUnitTempFolder();
-
-	public String getPowerShellExecutable() {
-		return System.getProperty("os.name").startsWith("Win") ? "powershell.exe" : "pwsh";
-	}
+public class TokenizerSensorTest extends BaseTest {
 
 	@Test
 	public void testIfFileIsSkipped() throws IOException {
